+++ CHANGELOG.md
@@ -1,13 +1,13 @@
 # CHANGELOG
 
-## Next Release
+## v0.9.2 (2023-09-02)
 
 - Fixes bugs where the request wasn't set correctly for parcels. users, or search
 - Fixes the input of creating a parcel from strings to numbers
 - Fixes the return types of various functions so 500s aren't thrown at runtime
 - Fixes a bug where an invalid ID passed to search would throw a 500 instead of returning an error to the user
 - Fixes a bug when creating a shipment where the variable `shipment` wasn't set
-- Temporarily downgrade to PHP 8.1 in Dockerfile
+- Fixes various bugs related to setup (names, version pins, etc)
 
 ## v0.9.1 (2023-02-24)